LPD-2447 Move to announcements module
