+++ modules/apps/announcements/announcements-web/src/main/java/com/liferay/announcements/web/internal/change/tracking/spi/display/AnnouncementsDeliveryCTDisplayRenderer.java
@@ -10,8 +10,22 @@ import com.liferay.change.tracking.spi.display.BaseCTDisplayRenderer;
 import com.liferay.change.tracking.spi.display.CTDisplayRenderer;
 import com.liferay.portal.kernel.exception.PortalException;
 import com.liferay.portal.kernel.language.Language;
+import com.liferay.portal.kernel.model.User;
+import com.liferay.portal.kernel.portlet.url.builder.PortletURLBuilder;
+import com.liferay.portal.kernel.security.permission.ActionKeys;
+import com.liferay.portal.kernel.service.UserLocalService;
+import com.liferay.portal.kernel.service.permission.UserPermissionUtil;
+import com.liferay.portal.kernel.theme.ThemeDisplay;
+import com.liferay.portal.kernel.util.Portal;
+import com.liferay.portal.kernel.util.WebKeys;
+import com.liferay.users.admin.constants.UserScreenNavigationEntryConstants;
+import com.liferay.users.admin.constants.UsersAdminPortletKeys;
 
 import java.util.Locale;
+
+import javax.portlet.PortletRequest;
+
+import javax.servlet.http.HttpServletRequest;
 
 import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
@@ -24,15 +38,52 @@ public class AnnouncementsDeliveryCTDisplayRenderer
 	extends BaseCTDisplayRenderer<AnnouncementsDelivery> {
 
 	@Override
+	public String getEditURL(
+		HttpServletRequest httpServletRequest,
+		AnnouncementsDelivery announcementsDelivery) {
+
+		ThemeDisplay themeDisplay =
+			(ThemeDisplay)httpServletRequest.getAttribute(
+				WebKeys.THEME_DISPLAY);
+
+		if (!UserPermissionUtil.contains(
+				themeDisplay.getPermissionChecker(),
+				announcementsDelivery.getUserId(), ActionKeys.UPDATE)) {
+
+			return null;
+		}
+
+		return PortletURLBuilder.create(
+			_portal.getControlPanelPortletURL(
+				httpServletRequest, UsersAdminPortletKeys.USERS_ADMIN,
+				PortletRequest.RENDER_PHASE)
+		).setMVCRenderCommandName(
+			"/users_admin/edit_user"
+		).setRedirect(
+			_portal.getCurrentURL(httpServletRequest)
+		).setParameter(
+			"p_u_i_d", announcementsDelivery.getUserId()
+		).setParameter(
+			"screenNavigationCategoryKey",
+			UserScreenNavigationEntryConstants.CATEGORY_KEY_PREFERENCES
+		).setParameter(
+			"screenNavigationEntryKey",
+			UserScreenNavigationEntryConstants.
+				ENTRY_KEY_ALERTS_AND_ANNOUNCEMENTS_DELIVERY
+		).buildString();
+	}
+
+	@Override
 	public Class<AnnouncementsDelivery> getModelClass() {
 		return AnnouncementsDelivery.class;
 	}
 
 	@Override
-	public String getTitle(Locale locale, AnnouncementsDelivery model)
+	public String getTitle(
+			Locale locale, AnnouncementsDelivery announcementsDelivery)
 		throws PortalException {
 
-		return _language.get(locale, model.getType());
+		return _language.get(locale, announcementsDelivery.getType());
 	}
 
 	@Override
@@ -56,11 +107,13 @@ public class AnnouncementsDeliveryCTDisplayRenderer
 		AnnouncementsDelivery announcementsDelivery = displayBuilder.getModel();
 
 		displayBuilder.display(
-			"delivery-id", announcementsDelivery.getDeliveryId()
-		).display(
-			"company-id", announcementsDelivery.getCompanyId()
-		).display(
-			"user-id", announcementsDelivery.getUserId()
+			"user",
+			() -> {
+				User user = _userLocalService.getUser(
+					announcementsDelivery.getUserId());
+
+				return user.getFullName();
+			}
 		).display(
 			"type", announcementsDelivery.getType()
 		).display(
@@ -74,5 +127,11 @@ public class AnnouncementsDeliveryCTDisplayRenderer
 
 	@Reference
 	private Language _language;
+
+	@Reference
+	private Portal _portal;
+
+	@Reference
+	private UserLocalService _userLocalService;
 
 }
