+++ modules/apps/announcements/announcements-web/src/main/java/com/liferay/announcements/web/internal/change/tracking/spi/reference/AnnouncementsDeliveryTableReferenceDefinition.java
@@ -1,13 +1,12 @@
 /**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
+ * SPDX-FileCopyrightText: (c) 2024 Liferay, Inc. https://liferay.com
  * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
  */
 
-package com.liferay.change.tracking.internal.spi.reference;
+package com.liferay.announcements.web.internal.change.tracking.spi.reference;
 
-import com.liferay.announcements.kernel.model.AnnouncementsEntry;
-import com.liferay.announcements.kernel.model.AnnouncementsEntryTable;
-import com.liferay.announcements.kernel.service.persistence.AnnouncementsEntryPersistence;
+import com.liferay.announcements.kernel.model.AnnouncementsDeliveryTable;
+import com.liferay.announcements.kernel.service.persistence.AnnouncementsDeliveryPersistence;
 import com.liferay.change.tracking.spi.reference.TableReferenceDefinition;
 import com.liferay.change.tracking.spi.reference.builder.ChildTableReferenceInfoBuilder;
 import com.liferay.change.tracking.spi.reference.builder.ParentTableReferenceInfoBuilder;
@@ -18,42 +17,39 @@ import org.osgi.service.component.annotations.Component;
 import org.osgi.service.component.annotations.Reference;
 
 /**
- * @author Preston Crary
+ * @author Brooke Dalton
  */
 @Component(service = TableReferenceDefinition.class)
-public class AnnouncementsEntryTableReferenceDefinition
-	implements TableReferenceDefinition<AnnouncementsEntryTable> {
+public class AnnouncementsDeliveryTableReferenceDefinition
+	implements TableReferenceDefinition<AnnouncementsDeliveryTable> {
 
 	@Override
 	public void defineChildTableReferences(
-		ChildTableReferenceInfoBuilder<AnnouncementsEntryTable>
+		ChildTableReferenceInfoBuilder<AnnouncementsDeliveryTable>
 			childTableReferenceInfoBuilder) {
-
-		childTableReferenceInfoBuilder.resourcePermissionReference(
-			AnnouncementsEntryTable.INSTANCE.entryId, AnnouncementsEntry.class);
 	}
 
 	@Override
 	public void defineParentTableReferences(
-		ParentTableReferenceInfoBuilder<AnnouncementsEntryTable>
+		ParentTableReferenceInfoBuilder<AnnouncementsDeliveryTable>
 			parentTableReferenceInfoBuilder) {
 
 		parentTableReferenceInfoBuilder.singleColumnReference(
-			AnnouncementsEntryTable.INSTANCE.companyId,
+			AnnouncementsDeliveryTable.INSTANCE.companyId,
 			CompanyTable.INSTANCE.companyId);
 	}
 
 	@Override
 	public BasePersistence<?> getBasePersistence() {
-		return _announcementsEntryPersistence;
+		return _announcementsDeliveryPersistence;
 	}
 
 	@Override
-	public AnnouncementsEntryTable getTable() {
-		return AnnouncementsEntryTable.INSTANCE;
+	public AnnouncementsDeliveryTable getTable() {
+		return AnnouncementsDeliveryTable.INSTANCE;
 	}
 
 	@Reference
-	private AnnouncementsEntryPersistence _announcementsEntryPersistence;
+	private AnnouncementsDeliveryPersistence _announcementsDeliveryPersistence;
 
 }
